# api/swagger/swagger.yaml
openapi: 3.0.0
info:
  version: "2.5.0"
  title: CCXT REST
  contact:
    name: "Adroit"
    email: "hello@adroit.ph"
    url: "https://adroit.ph/ccxt-rest-contact-us/"
  license:
    name: "MIT"
    url: "https://github.com/ccxt-rest/ccxt-rest/blob/master/LICENSE.txt"
servers:
  - url: "http://localhost:3000/"
...
paths:
...
  /exchange/{exchangeName}:
    get:
      description: Retreives the current exchange connection details given the {exchangeName} and access token in the header
      security:
        - bearerAuth: []
      tags:
        - Authentication API
      x-swagger-router-controller: exchange
      operationId: getConnection
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeResponse"
        "404":
          description: Exchange with that name is NOT supported
        "500":
          description: If an unexpected error occurred
    post:
      description: Creates a private connection to the exchange referenced in {exchangeName}
      security:
        - bearerAuth: []
      tags:
        - Authentication API
      operationId: createPrivateConnection
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
      requestBody:
        description: The exchange to create.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/exchangeConfig"
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessToken"
        "404":
          description: Exchange with that name is NOT supported
        "500":
          description: If an unexpected error occurred
        "503":
          description: Support for exchange is currently broken
      x-swagger-router-controller: exchange
    delete:
      description: Delete the exchange connection referenced by access token in the header
      security:
        - bearerAuth: []
      tags:
        - Authentication API
      x-swagger-router-controller: exchange
      operationId: deletePrivateConnection
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeResponse"
        "404":
          description: Exchange with that name is NOT supported
        "500":
          description: If an unexpected error occurred
  /exchange/{exchangeName}/fetchFundingRates:
    get:
      description: Data on the current, most recent, and next funding rates
      tags:
        - Public Data API
      x-swagger-router-controller: exchange
      operationId: fetchFundingRates
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
        - $ref: "#/components/parameters/optionalExchangeSymbolParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FundingRateResponse"
                x-content-type: application/json
        "500":
          description: If an unexpected error occurred
  /exchange/{exchangeName}/fetchFundingRate:
    get:
      description: Data on the current, most recent, and next funding rates
      tags:
        - Public Data API
      x-swagger-router-controller: exchange
      operationId: fetchFundingRate
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
        - $ref: "#/components/parameters/optionalExchangeSymbolParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FundingRateResponse"
                x-content-type: application/json
        "500":
          description: If an unexpected error occurred
  /exchange/{exchangeName}/markets:
    get:
      description: >-
        Get the markets of the exchange referenced by the {exchangeName}.


        <br/> *Parameters listed here are common to all exchanges. But any other parameter passed would be forwarded as well into the exchange.*
      security:
        - bearerAuth: []
      tags:
        - Public Data API
      x-swagger-router-controller: exchange
      operationId: markets
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketResponse"
                x-content-type: application/json
        "400":
          description: If the exchange itself complained about the parameters passed
        "401":
          description: If the exchange integration requires api key and secret for this function
        "403":
          description: If the exchange integration had an authentication issue (most probably nonce error)
        "404":
          description: Exchange with that name is NOT supported
        "500":
          description: If an unexpected error occurred
        "501":
          description: If the exchange integration does NOT support this function
        "504":
          description: If the exchange itself could not be reached because of some network error
  /exchange/{exchangeName}/orderBook:
    get:
      description: >-
        Get the order book of the exchange referenced by the {exchangeName} and `?symbol=`. 


        <br/> *Parameters listed here are common to all exchanges. But any other parameter passed would be forwarded as well into the exchange.*
      security:
        - bearerAuth: []
      tags:
        - Public Data API
      x-swagger-router-controller: exchange
      operationId: orderBook
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
        - $ref: "#/components/parameters/requiredExchangeSymbolParam"
        - name: limit
          in: query
          description: The limit of the exchange's order book to be retrieved.
          required: false
          style: form
          explode: true
          schema:
            type: number
        - name: exchangeSpecificParams
          in: query
          description: Any exchange specific parameters you want to pass in
          schema:
            type: object
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderBookResponse"
                x-content-type: application/json
        "400":
          description: If the exchange itself complained about the parameters passed
        "401":
          description: If the exchange integration requires api key and secret for this function
        "403":
          description: If the exchange integration had an authentication issue (most probably nonce error)
        "404":
          description: Exchange with that name is NOT supported
        "500":
          description: If an unexpected error occurred
        "501":
          description: If the exchange integration does NOT support this function
        "504":
          description: If the exchange itself could not be reached because of some network error
  /exchange/{exchangeName}/l2OrderBook:
    get:
      description: >-
        Get the Level 2 Order Book of the exchange referenced by the {exchangeName} and `?symbol=`.


        <br/> *Parameters listed here are common to all exchanges. But any other parameter passed would be forwarded as well into the exchange.*
      security:
        - bearerAuth: []
      tags:
        - Public Data API
      x-swagger-router-controller: exchange
      operationId: l2OrderBook
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
        - $ref: "#/components/parameters/requiredExchangeSymbolParam"
        - name: limit
          in: query
          description: The limit of the exchange's order book to be retrieved.
          required: false
          style: form
          explode: true
          schema:
            type: number
        - name: exchangeSpecificParams
          in: query
          description: Any exchange specific parameters you want to pass in
          schema:
            type: object
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderBookResponse"
                x-content-type: application/json
        "400":
          description: If the exchange itself complained about the parameters passed
        "401":
          description: If the exchange integration requires api key and secret for this function
        "403":
          description: If the exchange integration had an authentication issue (most probably nonce error)
        "404":
          description: Exchange with that name is NOT supported
        "500":
          description: If an unexpected error occurred
        "501":
          description: If the exchange integration does NOT support this function
        "504":
          description: If the exchange itself could not be reached because of some network error
...
  /exchange/{exchangeName}/tickers:
    get:
      description: >-
        Get the tickers of the exchange referenced by the {exchangeName} and `?symbol=`.


        <br/> *Parameters listed here are common to all exchanges. But any other parameter passed would be forwarded as well into the exchange.*
      security:
        - bearerAuth: []
      tags:
        - Public Data API
      x-swagger-router-controller: exchange
      operationId: tickers
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
        - $ref: "#/components/parameters/optionalExchangeSymbolParam"
        - name: exchangeSpecificParams
          in: query
          description: Any exchange specific parameters you want to pass in
          schema:
            type: object
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TickerResponse"
                x-content-type: application/json
        "400":
          description: If the exchange itself complained about the parameters passed
        "401":
          description: If the exchange integration requires api key and secret for this function
        "403":
          description: If the exchange integration had an authentication issue (most probably nonce error)
        "404":
          description: Exchange with that name is NOT supported
        "500":
          description: If an unexpected error occurred
        "501":
          description: If the exchange integration does NOT support this function
        "504":
          description: If the exchange itself could not be reached because of some network error
  /exchange/{exchangeName}/balances:
    get:
      description: >-
        Get the balances of the exchange referenced by the {exchangeName}. 


        <br/> *Parameters listed here are common to all exchanges. But any other parameter passed would be forwarded as well into the exchange.*
      security:
        - bearerAuth: []
      tags:
        - Private Data API
      x-swagger-router-controller: exchange
      operationId: balances
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
        - name: exchangeSpecificParams
          in: query
          description: Any exchange specific parameters you want to pass in
          schema:
            type: object
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceResponse"
        "400":
          description: If the exchange itself complained about the parameters passed
        "401":
          description: If the exchange integration requires api key and secret for this function
        "403":
          description: If the exchange integration had an authentication issue (most probably nonce error)
        "404":
          description: Exchange with that name is NOT supported
        "500":
          description: If an unexpected error occurred
        "501":
          description: If the exchange integration does NOT support this function
        "504":
          description: If the exchange itself could not be reached because of some network error
  /exchange/{exchangeName}/orders:
    get:
      description: >-
        Get the orders of the exchange referenced by the {exchangeName}. 


        <br/> *Parameters listed here are common to all exchanges. But any other parameter passed would be forwarded as well into the exchange.*
      security:
        - bearerAuth: []
      tags:
        - Private Data API
      x-swagger-router-controller: exchange
      operationId: fetchOrders
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
        - $ref: "#/components/parameters/optionalExchangeSymbolParam"
        - name: since
          in: query
          description: Retrieve the orders starting from 'since'
          required: false
          style: form
          explode: true
          schema:
            type: string
        - name: limit
          in: query
          description: The limit of the exchange's orders to be retrieved.
          required: false
          style: form
          explode: true
          schema:
            type: number
        - name: exchangeSpecificParams
          in: query
          description: Any exchange specific parameters you want to pass in
          schema:
            type: object
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
                x-content-type: application/json
        "400":
          description: If the exchange itself complained about the parameters passed
        "401":
          description: If the exchange integration requires api key and secret for this function
        "403":
          description: If the exchange integration had an authentication issue (most probably nonce error)
        "404":
          description: Exchange with that name is NOT supported
        "500":
          description: If an unexpected error occurred
        "501":
          description: If the exchange integration does NOT support this function
        "504":
          description: If the exchange itself could not be reached because of some network error
  /exchange/{exchangeName}/orders/open:
    get:
      description: >-
        Get the open orders of the exchange referenced by the {exchangeName}. 


        <br/> *Parameters listed here are common to all exchanges. But any other parameter passed would be forwarded as well into the exchange.*
      security:
        - bearerAuth: []
      tags:
        - Private Data API
      x-swagger-router-controller: exchange
      operationId: fetchOpenOrders
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
        - $ref: "#/components/parameters/optionalExchangeSymbolParam"
        - name: since
          in: query
          description: Retrieve the orders starting from 'since'
          required: false
          style: form
          explode: true
          schema:
            type: string
        - name: limit
          in: query
          description: The limit of the exchange's orders to be retrieved.
          required: false
          style: form
          explode: true
          schema:
            type: number
        - name: exchangeSpecificParams
          in: query
          description: Any exchange specific parameters you want to pass in
          schema:
            type: object
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
                x-content-type: application/json
        "400":
          description: If the exchange itself complained about the parameters passed
        "401":
          description: If the exchange integration requires api key and secret for this function
        "403":
          description: If the exchange integration had an authentication issue (most probably nonce error)
        "404":
          description: Exchange with that name is NOT supported
        "500":
          description: If an unexpected error occurred
        "501":
          description: If the exchange integration does NOT support this function
        "504":
          description: If the exchange itself could not be reached because of some network error
  /exchange/{exchangeName}/orders/closed:
    get:
      description: >-
        Get the closed orders of the exchange referenced by the {exchangeName}.


        <br/> *Parameters listed here are common to all exchanges. But any other parameter passed would be forwarded as well into the exchange.*
      security:
        - bearerAuth: []
      tags:
        - Private Data API
      x-swagger-router-controller: exchange
      operationId: fetchClosedOrders
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
        - $ref: "#/components/parameters/optionalExchangeSymbolParam"
        - name: since
          in: query
          description: Retrieve the orders starting from 'since'
          required: false
          style: form
          explode: true
          schema:
            type: string
        - name: limit
          in: query
          description: The limit of the exchange's orders to be retrieved.
          required: false
          style: form
          explode: true
          schema:
            type: number
        - name: exchangeSpecificParams
          in: query
          description: Any exchange specific parameters you want to pass in
          schema:
            type: object
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
                x-content-type: application/json
        "400":
          description: If the exchange itself complained about the parameters passed
        "401":
          description: If the exchange integration requires api key and secret for this function
        "403":
          description: If the exchange integration had an authentication issue (most probably nonce error)
        "404":
          description: Exchange with that name is NOT supported
        "500":
          description: If an unexpected error occurred
        "501":
          description: If the exchange integration does NOT support this function
        "504":
          description: If the exchange itself could not be reached because of some network error
  /exchange/{exchangeName}/trades/mine:
    get:
      description: >-
        Get my trades of the exchange referenced by the {exchangeName}. 


        <br/> *Parameters listed here are common to all exchanges. But any other parameter passed would be forwarded as well into the exchange.*
      security:
        - bearerAuth: []
      tags:
        - Private Data API
      x-swagger-router-controller: exchange
      operationId: fetchMyTrades
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
        - $ref: "#/components/parameters/optionalExchangeSymbolParam"
        - name: since
          in: query
          description: Retrieve the trades starting from 'since'
          required: false
          style: form
          explode: true
          schema:
            type: string
        - name: limit
          in: query
          description: The limit of the exchange's trades to be retrieved.
          required: false
          style: form
          explode: true
          schema:
            type: number
        - name: exchangeSpecificParams
          in: query
          description: Any exchange specific parameters you want to pass in
          schema:
            type: object
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradeResponse"
                x-content-type: application/json
        "400":
          description: If the exchange itself complained about the parameters passed
        "401":
          description: If the exchange integration requires api key and secret for this function
        "403":
          description: If the exchange integration had an authentication issue (most probably nonce error)
        "404":
          description: Exchange with that name is NOT supported
        "500":
          description: If an unexpected error occurred
        "501":
          description: If the exchange integration does NOT support this function
        "504":
          description: If the exchange itself could not be reached because of some network error
...
components:
  schemas:
...
    AccessToken:
      required:
        - token
      properties:
        token:
          type: string
          description: >-
            The JWT token that needs to be added into the 'Authorization'
            header with the 'Bearer ' prefix. For example, given a token of xyz,
            do a secured request with 'Authorization: Bearer xyz'
    ExchangeResponse:
      required:
        - id
        - countries
        - enableRateLimit
        - name
        - private
      type: object
      properties:
        id:
          type: string
          description: >-
            The id of the exchange. When you created the exchange (via
            POST:/exchanges/{exhangeName}), the 'id' parameter there becomes the
            name here
        name:
          type: string
          description: >-
            The name of the exchange.
        private:
          type: boolean
          description: Whether this exchange is private (has apiKey) or public (no apiKey)
        enableRateLimit:
          type: boolean
          description: >-
            Whether to enable the built in rate limiter or not. The built in
            rate limiter is an approximation of the actual exchange's limit. To
            have a more accurate rate limiting, set this to false and implement
            the rate limiter on your client
          default: true
        countries:
          type: array
          description: The list of countries where this exchange is a member of
          items:
            type: string
        rateLimit:
          type: integer
          description: >-
            A request rate limit in milliseconds. Specifies the required minimal
            delay between two consequent HTTP requests to the same exchange. If
            enableRateLimit is set to false, this would be ignored.
        twofa:
          type: boolean
          description: Whether to enable two factor authentication or not
          default: false
        has:
          $ref: "#/components/schemas/ExchangeHasCapabilities"
        urls:
          type: object
          properties: {}
          description: Collection of URLs this exchange has
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    exchangeNameParam:
      name: exchangeName
      in: path
      description: >-
        The name of the exchange. Possible values are any of the result of
        GET:/exchanges.
      required: true
      style: simple
      explode: false
      schema:
        $ref: "#/components/schemas/Exchange"